ADITI: Affordable Diagnostics Enabled Thermal Incubator.
It uses a 25V N-MOSFET to drive a LiTex heating textile at 9V using PWM
(PID heating to be added)
Connections:
1)Connect SDA, SCL of MAX30100 to D0,D1 respectively and Vin,Gnd to 3.3V
and Gnd.
2)Connect the analog output of LM35 to A1 pin and power and Gnd to 5V
and Gnd.
3)Connect one terminal of eeonyx to 5V and the other to A0, Connect a
6.8kohm resistor and 10uF filter cap between A1 and Gnd
4)Connect one terminal of the LiTex heating pad to the Positive terminal
of the XL6009 Boost converter and the negative of heating pad to the
drain of the IRL3713 NMOSFET
and connect source of MOSFET to the negative terminal of boost
convertor. Connect gate to PWM pin D3.
5)Connect the ADS1292 ECG sheild

DISCLAIMER **While I aim to make this more safe, accurate it this is
experimental software (and pre-beta) at that.
It is not suitable for any particular purpose. No life-critical devices
should be based on this software.**
Work by Vignesh Ravichandran (hello@rvignesh.xyz)